+++ src/index.html
@@ -5,19 +5,14 @@
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
   <title>To-Do List</title>
   <link rel="stylesheet" href="style.css">
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.15.3/css/all.min.css">
 </head>
 <body>
   <div class="container">
     <h1>To-Do List</h1>
-    <div class="add-item">
-      <input type="text" id="newItemInput" placeholder="Add to your list...">
-    </div>
     <ul id="taskList"></ul>
-    <button id="clearCompletedButton">Clear Completed</button>
+    <button class="clear-completed-button">Clear Completed</button>
   </div>
 
   <script src="index.js" type="module"></script>
 </body>
-</html>
 </html>
