+++ src/style.css
@@ -1,3 +1,7 @@
+body {
+  font-family: Arial, sans-serif;
+}
+
 .container {
   max-width: 400px;
   margin: 0 auto;
@@ -9,16 +13,7 @@
 
 h1 {
   text-align: center;
-}
-
-.add-item {
-  display: flex;
-  margin-bottom: 10px;
-}
-
-.add-item input {
-  flex-grow: 1;
-  padding: 5px;
+  margin-bottom: 20px;
 }
 
 ul {
@@ -29,27 +24,26 @@ ul {
 li {
   display: flex;
   align-items: center;
-  margin-bottom: 5px;
-  padding: 5px;
+  justify-content: space-between;
+  margin-bottom: 10px;
+  padding: 10px;
   background-color: #fff;
   border: 1px solid #ccc;
   border-radius: 3px;
+  box-shadow: 0 2px 4px rgba(0, 0, 0, 0.1);
 }
 
 li.completed {
   text-decoration: line-through;
 }
 
-li span {
-  flex-grow: 1;
-}
-
-.remove-item {
+.clear-completed-button {
+  display: block;
+  margin-top: 10px;
+  padding: 5px 10px;
+  background-color: #2689a741;
+  color: black;
+  border: none;
+  border-radius: 3px;
   cursor: pointer;
-  margin-left: 10px;
-  color: rgb(13, 13, 13);
-}
-
-li input[type="checkbox"] {
-  margin-right: 10px;
 }
